rename `validation_road.py` to `road.py` and move up to root directory
