--- road/config.toml
@@ -1,70 +1,0 @@
-[LOADED_NETWORK]
-path = "Q:/Model Development/Simwrapper Validation/notebooks/roads/Data"
-
-# The Loaded_network file name is LOAD[TOD].dbf by default
-timeperiods = ["AM", "MD", "PM", "EV", "EA"]
-
-# The needed CHAMP output columns
-columns = ["A", "B", "AT", "FT", "V_1"]
-
-[OBSERVED_COUNTS]
-# The observed data file path
-obs_filepath = "X:/Projects/Miscellaneous/validation_simwrapper/roads2023/observed_counts.csv"
-
-# The columns used for calculation, usually are 'A,B, and Counts'
-obs_usecols = ["A", "B", "Daily", "AM", "MD", "PM", "EV", "EA"]
-
-# The columns we want to include from the observed data, DONT CHANGE A, B!
-obs_extra_columns = ["A", "B", "Loc Type", "Source", "Source ID", "Location", "Dir"]
-
-[SCATTER_INPUT]
-# The time period we want to display in the scatter plot
-chosen_period = "Daily"
-
-# Change the variables to show different group_by result
-classification_col = "Loc Type"
-
-# cols displayed and information we want in our generated csv
-combined_df_cols = ["Loc Type", "A", "B", "AT", "FT", "Source ID", "AT Group", "FT Group", "Location", "Dir"]
-
-output_file_name = "scatter.csv"
-
-[EST_SCATTER_PLOT]
-xfield = "Observed Volume"
-yfield = "Estimated Volume"
-
-# Information we want to display in the hover over
-fields = ["Observed Volume", "Estimated Volume", "Loc Type", "A", "B", "Location", "Dir"]
-
-# Be careful to state all nominal fields from the above fields
-nominal_fields = ["Loc Type", "A", "B", "Location", "Dir"]
-name = "Estimation Volume"
-est_output_template = "{classification_col_types}_{name}_est.vega.json"
-
-[PERCENT_SCATTER_PLOT]
-xfield = "Observed Volume"
-yfield = "Percent Errors"
-fields = ["Observed Volume", "Percent Errors", "Loc Type", "A", "B", "Location",  "Dir"]
-nominal_fields = ["Loc Type", "A", "B", "Location", "Dir"]
-name = "Percent Errors"
-diffpercent_output_template = "{classification_col_types}_{name}_diffpercent.vega.json"
-
-dashboard_number = 1
-
-[STATS_INPUT]
-# cols displayed
-combined_df_cols = ["Loc Type", "A", "B", "AT", "FT", "Source ID", "AT Group", "FT Group", "Location",  "Dir"]
-# The output csv file name
-output_file_name = 'scatter.csv'
-
-[MAP_INPUT]
-freeflow_dir = "X:/Projects/DTX/CaltrainValidation/s8_2019_Base/freeflow.shp"
-shp_out_dir = "X:/Projects/Miscellaneous/validation_simwrapper/roads2023/map"
-output_filename = 'map.csv'
-
-
-[AT]
-values = ["Core/CBD", "Core/CBD", "UrbBiz", "Urb", "Sub", "Sub"]
-
-[FT]
-values = ["Fwy/Ramp", "Fwy/Ramp", "Fwy/Ramp", "Col", "Fwy/Ramp", "", "Art", "", "Loc", "", "Loc", "Art", "", "", "Art"]
